feat: read delegators data from file

+++ src/main.ts
@@ -1,4 +1,4 @@
-import { existsSync, mkdirSync, readFileSync, rmdirSync, writeFileSync } from "fs";
+import { existsSync, mkdirSync, readFile, readFileSync, rmdirSync, writeFileSync } from "fs";
 import { MpDaoVoteContract, VoterInfo } from "./contracts/mpdao-vote";
 import { setRpcUrl, yton } from "near-api-lite";
 import { argv, cwd, env } from "process";
@@ -769,10 +769,14 @@ async function getENOsStakeDataAndInsertIt(contracts?: string[]) {
     }
 
     const contractsToAdd = contracts || getENOsContracts()
-    const data = await generateDelegatorTableDataSince(startUnixTimestamp, endUnixTimestamp, contractsToAdd)
-    const now = Date.now()
-    const delegatorTableFileName = `ENOs/temp_delegators_table_data_${now}.json`
-    writeFileSync(delegatorTableFileName, JSON.stringify(data))
+    const delegatorTableFileName = `ENOs/temp_delegators_table_data.json`
+    let data
+    if(existsSync(delegatorTableFileName)) {
+        data = JSON.parse(readFileSync(delegatorTableFileName, 'utf-8'))
+    } else {
+        data = await generateDelegatorTableDataSince(startUnixTimestamp, endUnixTimestamp, contractsToAdd)
+        writeFileSync(delegatorTableFileName, JSON.stringify(data))
+    }
     if (data.length > 0) {
         const isSuccess = await insertENOsData(data)
         if (isSuccess && !contracts) { // If contract is provided, we don't want to update, since all the other contracts may have not been updated yet
@@ -785,9 +789,14 @@ async function getENOsStakeDataAndInsertIt(contracts?: string[]) {
         }
     }
 
-    const dataByDelegator = await generateTableDataByDelegatorSince(startUnixTimestampByDelegator, endUnixTimestampByDelegator, contractsToAdd)
-    const dataByDelegatorFileName = `ENOs/temp_data_by_delegators_${now}.json`
-    writeFileSync(dataByDelegatorFileName, JSON.stringify(dataByDelegator))
+    let dataByDelegator
+    const dataByDelegatorFileName = `ENOs/temp_data_by_delegators.json`
+    if(existsSync(dataByDelegatorFileName)) {
+        dataByDelegator = JSON.parse(readFileSync(dataByDelegatorFileName, 'utf-8'))
+    } else {
+        dataByDelegator = await generateTableDataByDelegatorSince(startUnixTimestamp, endUnixTimestamp, contractsToAdd)
+        writeFileSync(dataByDelegatorFileName, JSON.stringify(dataByDelegator))
+    }
     if (dataByDelegator.length > 0) {
         const isSuccess = await insertENOsByDelegatorData(dataByDelegator)
         if (isSuccess && !contracts) {// If contract is provided, we don't want to update, since all the other contracts may have not been updated yet
